Store selected scan in generic chromatogram selection.
